backend/requirements.txt
fastapi[standard]>=0.112.0,<0.113.0
requests==2.31.0
keyboard
python-decouple
matplotlib
pandas
seaborn
uvicorn
httpx
pillow
mss
bs4
cachetools
opencv-python 
easyocr
fuzzywuzzy
python-Levenshtein
torch
transformers